removed index.js and renamed main.jsx to index.jsx; moved nav and footer components into root.jsx
